+++ NEWS.md
@@ -5,5 +5,6 @@
 -   Added options management. `cdo_options_use()` will use options for an operation. `cdo_option_set()` will set global options to use for all operations. `cdo_options_clear()` will clear the global options.
 -   The new `cdo_operator()` allows the user to define an arbitrary operator in case the some operator is missing in the package. Operators are run with the `cdo()` function.
 -   Added operators with zero outputs.
--   By default, temporay files output are deleted when not accessible from the R session.
+-   By default, temporary files output are deleted when not accessible from the R session.
 -   Adds change operators.
+-   Adds `cdo_execute_list()` to execute a list of cdo operations. This is the preferred way of applying operations to a set of files since it preserves ephemeral files correctly.